implementation of the framework and bechmarks

+++ caching/cached_dict.py
@@ -12,7 +12,8 @@ class CachedTensorDict:
     DATASET_NAME = 'data'
     TENSOR_FILE = 'values.h5py'
     INDEX_FILE = 'index.pkl'
-    def __init__(self, path: str, name:str, shape: Tuple[int, ...], write_size: int = 1000):
+
+    def __init__(self, path: str, name: str, shape: Tuple[int, ...], write_size: int = 1000, read_only: bool = False):
         assert os.path.isdir(path)
 
         full_path = os.path.join(path, name)
@@ -24,41 +25,51 @@ class CachedTensorDict:
         self.index_file = os.path.join(full_path, self.INDEX_FILE)
 
         self.write_size = write_size
+        self.read_only = read_only
         self._base_shape = shape
 
-        if not os.path.isfile(self.index_file):
-            with h5py.File(self.tensor_file, 'w') as f:
-                f.create_dataset(
-                    self.DATASET_NAME,
-                    shape=(0, *shape),
-                    maxshape=(None, *shape)
-                )
-            with open(self.index_file, 'wb') as file:
-                pickle.dump({}, file)
+        print(f"Opening {self.tensor_file}")
 
-        self.f = h5py.File(self.tensor_file, 'r')
-        self._cache = self.f[self.DATASET_NAME]
+        try:
+            if self.read_only:
+                f = h5py.File(self.tensor_file, 'r')
+            else:
+                f = h5py.File(self.tensor_file, 'a')
+                if not self.DATASET_NAME in f:
+                    f.create_dataset(
+                        self.DATASET_NAME,
+                        shape=(0, *shape),
+                        maxshape=(None, *shape)
+                    )
+                if not os.path.isfile(self.index_file):
+                    with open(self.index_file, 'wb') as file:
+                        pickle.dump({}, file)
+        except BlockingIOError as e:
+            print(f"Unable to open {self.tensor_file}")
+            raise e
+
+        assert self.DATASET_NAME in f
+        self._cache = f[self.DATASET_NAME]
 
         with open(self.index_file, 'rb') as file:
             self._index = pickle.load(file)
+
         self._new_cache = {}
+        self._f = f
 
     @property
     def shape(self) -> Tuple[int, ...]:
         return (len(self._index), *self._base_shape)
 
     def write_cache_to_disk(self):
+        assert not self.read_only
+
         if len(self._new_cache) == 0:
             return
 
         # Write the freshly added data
         old_cache_size = self._cache.shape[0]
         new_cache_size = len(self._index)
-
-        # Change open mode
-        self.f.close()
-        self.f = h5py.File(self.tensor_file, 'w')
-        self._cache = self.f[self.DATASET_NAME]
 
         # Resize to fit the new data
         self._cache.resize((new_cache_size, *self._base_shape))
@@ -71,30 +82,22 @@ class CachedTensorDict:
                 assert idx >= old_cache_size, f"{idx}, {old_cache_size}"
                 self._cache[idx] = value
 
-        # Change back to read mode
-        self.f.close()
-        self.f = h5py.File(self.tensor_file, 'r')
-        self._cache = self.f[self.DATASET_NAME]
-
         # Update the index
         with open(self.index_file, 'wb') as f:
             pickle.dump(self._index, f)
 
         self._new_cache = {}
 
-    def _add_to_cache(self, key: str, value: Optional[np.ndarray]):
+    def _add_to_cache(self, key: str, value: np.ndarray):
+        assert not self.read_only
         assert not (key in self._index) and not (key in self._new_cache)
 
-        if value is None:
-            self._index[key] = -1
-        else:
-            self._index[key] = len(self._index)
+        self._index[key] = len(self._index)
         self._new_cache[key] = value
 
         # Write to disk when the data in memory is too much
-        if len(self._index)-len(self._cache) >= self.write_size:
+        if len(self._index) - self._cache.shape[0] >= self.write_size:
             self.write_cache_to_disk()
-
 
     def __contains__(self, key: str):
         return key in self._index or key in self._new_cache
@@ -108,36 +111,52 @@ class CachedTensorDict:
                 value = self._cache[idx]
             return value
         elif isinstance(key, list):
-            # Write everithing to cache
-            self.write_cache_to_disk()
+            # Write everything to cache
+            if not self.read_only:
+                self.write_cache_to_disk()
 
             # Then look up
             ids = np.array([self._index[k] for k in key])
-            if len(ids) > 1:
-                # Make sure the ids are ordered
-                if not np.all(ids[:-1] <= ids[1:]):
-                    # Step 1: Sort the IDs and remember the original indices
-                    id_order = np.argsort(ids)
-                    sorted_ids = ids[id_order]
 
-                    # Step 2: Access the array using the sorted IDs
-                    values = self._cache[sorted_ids]
-
-                    # Step 3: Sort the values back to the original order
-                    inverse_indices = np.empty_like(ids)
-                    inverse_indices[id_order] = np.arange(len(key))
-                    values = values[inverse_indices]
-                else:
-                    values = self._cache[ids]
-            else:
+            # Make sure the ids are ordered (h5py supports only ordered access)
+            if np.all(ids[:-1] < ids[1:]):
                 values = self._cache[ids]
-            return values
 
+            # If the ids are not in order
+            else:
+                # Step 1: Sort the IDs and remember the original indices
+                id_order = np.argsort(ids)
+                sorted_ids = ids[id_order]
+
+                # Step 2: Check for duplicate ids and remove them from the query
+                duplicates = not np.all(sorted_ids[:-1] < sorted_ids[1:])
+                if duplicates:
+                    non_duplicate_mask = np.concatenate(
+                        [np.array([True]).reshape(1), sorted_ids[:-1] < sorted_ids[1:]], 0
+                    )
+                    sorted_ids = sorted_ids[non_duplicate_mask]
+                    id_lookup = np.cumsum(non_duplicate_mask) - 1
+                else:
+                    id_lookup = np.arange(id_order.shape[0])
+
+                # Step 3: Access the array using the sorted IDs
+                values = self._cache[sorted_ids]
+
+                # Step 4: Sort the values back to the original order
+                inverse_ids = np.empty_like(id_order)
+                inverse_ids[id_order] = id_lookup
+                values = values[inverse_ids]
+
+            return values
 
     def __len__(self) -> int:
         return len(self._index)
 
+    def __delitem__(self, key):
+        raise NotImplementedError("Item deletion is not implemented")
+
     def __setitem__(self, key: str, value: Optional[np.ndarray]):
+        assert not self.read_only
         assert value is None or value.shape == self._base_shape
 
         if not key in self:
@@ -149,8 +168,25 @@ class CachedTensorDict:
             self._cache[idx] = value
 
     def __del__(self):
-        self.write_cache_to_disk()
-        self.f.close()
+        self.close()
 
     def __bool__(self):
         return True
+
+    def flush(self):
+        if not self.read_only:
+            self.write_cache_to_disk()
+
+    def close(self):
+        self.flush()
+        print(f"Closing {self.tensor_file}")
+        self._f.close()
+
+    def keys(self):
+        return self._index.keys()
+
+    def __iter__(self):
+        return iter(self.keys())
+
+    def values(self):
+        return [self[key] for key in self]
